Extract models in `url_metadata` to a different file
