+++ datalad_registry/blueprints/api/url_metadata/__init__.py
@@ -2,10 +2,10 @@
 # i.e. the metadata of datasets at individual URLs.
 
 from flask_openapi3 import APIBlueprint, Tag
-from pydantic import BaseModel, Field, StrictStr
 
 from datalad_registry.models import URLMetadata, db
 
+from .models import PathParams, URLMetadataModel
 from .. import API_URL_PREFIX, COMMON_API_RESPONSES
 
 bp = APIBlueprint(
@@ -15,31 +15,6 @@ bp = APIBlueprint(
     abp_tags=[Tag(name="URL Metadata", description="API endpoints for URL metadata")],
     abp_responses=COMMON_API_RESPONSES,
 )
-
-
-class PathParams(BaseModel):
-    """
-    Pydantic model for representing the path parameters for the URL metadata API
-    endpoints.
-    """
-
-    url_metadata_id: int = Field(..., description="The ID of the URL metadata")
-
-
-class URLMetadataModel(BaseModel):
-    """
-    Model for representing the database model URLMetadata for communication
-    """
-
-    dataset_describe: StrictStr
-    dataset_version: StrictStr
-    extractor_name: StrictStr
-    extractor_version: StrictStr
-    extraction_parameter: dict
-    extracted_metadata: dict
-
-    class Config:
-        orm_mode = True
 
 
 @bp.get("/<int:url_metadata_id>", responses={"200": URLMetadataModel})
