+++ datalad_registry/blueprints/api/url_metadata/models.py
@@ -1,0 +1,26 @@
+from pydantic import BaseModel, Field, StrictStr
+
+
+class PathParams(BaseModel):
+    """
+    Pydantic model for representing the path parameters for the URL metadata API
+    endpoints.
+    """
+
+    url_metadata_id: int = Field(..., description="The ID of the URL metadata")
+
+
+class URLMetadataModel(BaseModel):
+    """
+    Model for representing the database model URLMetadata for communication
+    """
+
+    dataset_describe: StrictStr
+    dataset_version: StrictStr
+    extractor_name: StrictStr
+    extractor_version: StrictStr
+    extraction_parameter: dict
+    extracted_metadata: dict
+
+    class Config:
+        orm_mode = True
